extract remaining ogdl classes to separate files
